Clean up govuk roles a bit more
Remove govuk- prefix - these are already present in our account names,
so there's no need to show them in the role name too.

Replace administrators with admin - it's just as clear, and a lot
shorter.

The existing one is to remove the gross number (/172398498723) that AWS
puts on the end of federated logins - I don't know what this number is,
so it's definitely not useful to show it to me.